Get "dynamic <IPv6addr>" working for IPAM.
Similar to static_dynamic_ip, this adds a static_dynamic_ip6 field for
SwitchPort to indicate if there is a static IPv6 address configured for
a dynamic MAC. This is then used as part of the calcultion for if a
dynamic IPv6 address needs to be generated.

NOTE: THIS CODE IS CURRENTLY BROKEN

Currently when run, this results in a transaction failure. This results
in ovn-northd-ddlog being no longer functional.