Add initial Port Binding statistics (pinctrl counts)
This introduces centralized statistics gathering for Port Bindings. In
this initial patch, the only statistics collected are counts of types of
actions handled by pinctrl.

Port_Bindings now have a list of Port_Binding_Statistics added to them.
Each instance of Port_Binding_Statistics on the Port_Binding corresponds
to a specific chassis on which the statistic was captured.

Two ovn-sbctl commands are added which allow for the statistics to be
printed or cleared. The formatting of the output could stand to be
improved.

This patch is being presented as an RFC:

1) Is this sort of table layout appropriate for the task?
2) Are the operations added to pinctrl.c appropriate? In other words, do
   you foresee performance-related issues from the added
   statistics-gathering overhead? The current incarnation triggers an
   ovn-controller run for every added statistic, but it may be more
   appropriate to run every X seconds instead.
3) What other pinctrl-related statistics would be appropriate to add in
   a first patch? Currently, this patch adds only the number of times a
   certain action opcode was handled. There is no individual packet data
   collected, no count of successes vs. errors, no timestamps for when
   the actions were handled, etc.