pinctrl: Add Chassis MAC_Bindings for all router addresses.
Previous behavior of ovn-controller was to only create MAC bindings for
the same addresses for which we would send gARPs. That is:

* A logical switch has its "nat_addresses" configured.
* That logical switch has a localnet port.

This makes sense for gARPs: we only want to send gARPs for addresses
accessible from outside of OVN, and we only want to send gARPs for
addresses explicitly enabled by the administrator.

For directly adding MAC bindings, though, this makes less sense. In the
case where a VM or pod on the underlay network wishes to reach another
VM or pod via a DNAT or load balancer address, an ARP was required to
learn the MAC binding. When the information is all available in the
database, this seems like an unnecessary step.

This commit makes the following changes:
* MAC_Binding now has a "chassis_name" column. This is used to identify
  the chassis that added the MAC Binding IF the MAC Binding was auto-
  generated by ovn-controller. This is also used to ensure that stale
  bindings can be cleaned up.
* The term "local garp" has been replaced with "chassis MAC binding".
  Removing the word "garp" helps to remove any implication of packets
  being transmitted. The word "chassis" precedes "MAC binding" to
  indicate it is a MAC binding added by a chassis and with a
  chassis_name in its records. MAC bindings with no chassis_name were
  added dynamically due to reception of ARP/gARP packets.
* Chassis MAC Binding code has been separated from gARP code since they
  do not operate using the same sets of addresses.
* Chassis MAC Bindings are added for all router_addresses on a port
  binding. This means all NAT, Load Balancer, and router interface
  addresses are added for routers that are residents of the chassis.

Signed-off-by: Mark Michelson <mmichels@redhat.com>