northd: Allow "enable_router_port_conntrack" to work on LBs.
Prior to this commit, if enable_router_port_conntrack (or
enable_router_port_acl) were enabled, then requests would be load
balanced properly. However, replies would not be unDNATted. This is
because conntrack was skipped in the OUT_PRE_LB stage, resulting in
REGBIT_CONNTRACK_NAT not being set. Since this was not set, it meant
that the "ct_lb;" call in the OUT_PRE_STATEFUL stage would not happen.

This commit skips conntrack in the OUT_PRE_LB stage if the
enable_router_port_conntrack option is enabled. This allows the "ct_lb;"
action to execute in the OUT_PRE_STATEFUL stage, thus unDNATting the
reply.

Signed-off-by: Mark Michelson <mmichels@redhat.com>