Added debug to failing system test.
The system test is failing because the expected conntrack entry with the
ACL ID label is not present. There is a conntrack entry that sets the
mark bit to persist the connection. However, the expected label is not
there.

I've checked and the ACL ID is present in the SB DB. northd finds this
ID and creates a flow that sets the REG_ACL_ID register to 1. I can see
the OF flow in OVS as well. However, it appears to never have been hit.
It's confusing why there is a conntrack entry with the mark set as
expected, but the label is not set. If the flow was never hit, then I'd
expect the conntrack entry not to have the mark or label set.