Datapath_Binding: Separate type and UUID external-ids.
Prior to this commit, if you wanted to find the corresponding northbound
UUID for a southbound Logical Datapath, you could find it in one of two
places:

For logical switches, it was in external-ids:logical-switch.
For logical routers, it was in external-ids:logical-router.

With this commit, we are separating the type and UUID. The type is a new
column, and the datapath UUID is now propagated from the northbound logical
datapath. There had been concerns about this UUID propagation in the
past since it was possible for a logical switch and logical router to
have the same UUID. While this was technically true in the northbound
database, ovn-northd would flag this as an error and not allow the
duplicated UUID. Therefore, we are not breaking any existing
deployments with this change.

Note that there is a seemingly-unrelated change in the check packet
length test in tests/ovn.at. This test started failing because its use
of `grep` was picking up the new "nb_uuid" column accidentally. The
change to the test uses `fetch_column` since it is more precise.

The ovn-northd duplicated datapaths test has been removed from the
testsuite. The reason is that it is now impossible to create a
duplicated datapath since the OVSDB server will not allow us to insert a
datapath binding with a duplicated UUID.

Signed-off-by: Mark Michelson <mmichels@redhat.com>