inc-proc-eng: Change handler result to an Either type.
The incremental engine currently provides an optional callback for
engine nodes called "get_compute_failure_info()". The goal of this
callback is to assist in debugging why an engine node had to recompute.

There are flaws with this callback, though.

First, it is called on the input that caused the failure, rather than
the node where the failure was detected. The input has no knowledge
of why it caused a failure. The best that can be done is to dump all
data from the input. Depending on the amount of data in the node, this
could result in a massive dump of info, and trying to pinpoint where the
failure point is can be difficult.

Second, the callback is called at inconsistent times. If an input
handler fails, then the callback is called before the node recomputes.
If a node has no change handler for an input, then the callback is
called after the node recomputes. This can lead to confusion when
attempting to debug based on logs.

Third, since the callback is optional, very few nodes actually
implement it. The only nodes that implement this callback are the DB
nodes. All others result in a message about "missing compute failure
debug info."

To remedy this, this commit makes it so that change handlers must return
a string indicating the failure reason when returning. The way this is
done is by changing engine_input_handler_result from an enum to an
Either type. To ease the transition easier, the existing enum values have
been converted to macros that equate to the same thing. The EN_UNHANDLED
value now takes a format string as an argument in order to allow precise
reporting for an input handler's failure reason.

Since this commit is already on the large side, this patch does not
attempt to meaningfully return failure reasons in change handlers.
Instead, a placeholder string of "XXX FIXME" is used here. Commits
following this one will change these to proper failure reasons.

Signed-off-by: Mark Michelson <mmichels@redhat.com>