inc-proc-eng: Update "simple" XXX FIXMEs.
This commit changes the simplest EN_UNHANDLED reasons from "XXX FIXME"
to something more meaningful. In this case, "simple" means that there is
a direct cause that can be easily stated for causing the recompute to
occur.

The next commits will deal with more complex situations, like when an
indirect function is used to determine the cause for the recompute.

Signed-off-by: Mark Michelson <mmichels@redhat.com>