SB schema: Add nb_uuid column to Datapath_Binding.
Commit <insert SHA here> changed ovn-northd to sync the NB logical
datapath UUID with the corresponding SB Datapath_Binding. They thus
would have the same UUID. This causes issues on upgrades since it
involves having to essentially recreate all southbound datapath
bindings, as well as all records that refer to those southbound datapath
bindings.

This commit changes so that instead of having the same UUID, the
southbound datapath_binding has a new nb_uuid column that contains the
UUID of the northbound logical datapath. This way, there is still a
simple way to link the Datapath_Binding to the northbound logical
datapath, but there is no need to have to update pre-existing
Datapath_Bindings to have new UUIDs.

Signed-off-by: Mark Michelson <mmichels@redhat.com>