ovn-util: Add generic "Either" type macros.
C does not provide union types the same as other higher level languages.
You can't define, for instance, a function that can return an int when
things are successful and a string when things fail. In order to do
this, the two types need to be a union within a struct, with a boolean
field to distinguish which field of the union to use.

This commit provides general purpose macros that help to define Either
types, determine success or failure, and get/set success and failure
valures. These macros will be used in upcoming commits.

Signed-off-by: Mark Michelson <mmichels@redhat.com>