ovn-util: Add annotated_bool.
An annotated bool is a special case of an Either type structure. In
practice it works similar to a boolean, except that a "true" value is
represented with nothing, and a "false" value is represented with a
reason string.

The intended use for this is for functions that return true upon success
and false upon failure. Since functions usually may have multiple
reasons they may fail, using an annotated_bool allows for the failure
case to express a reason for the failure.

We will be using these in upcoming commits to better explain the reasons
for incremental engine handler failures.

Signed-off-by: Mark Michelson <mmichels@redhat.com>