Add sparse array.
This operates as an alternate dynamic array type from vector.

Vectors are useful for allocating a contiguous block of elements.
Inserting an element will cause other elements to shift, resulting in a
still-contiguous block of elements. Removing an element causes elements
above it to be shifted down to fill in the gap. Attempting to insert an
element past the bounds of the array will not work.

Sparse arrays serve a different purpose. Their goal is to ensure that
elements maintain their index as long as they are in the array.
Inserting an element causes it to take the lowest free slot in the
array. If an element is inserted at a specific index, then it replaces
the element that was there previously. If an element is removed, it
results in a hole in the array. If you attempt to add an element past
the bounds of the array, then the array expands to allow it.

This is all technically possible with vectors, too. There is code in
northd.c that uses vector_get() cleverly to avoid outright removal of
elements. However, that completely breaks any attempted traversal of the
vector, and the length reported by the vector is no longer correct. In
my personal opinion, this is a by-the-book definition of a "code smell".

Commits following this one will start making use of sparse array.

Signed-off-by: Mark Michelson <mmichels@redhat.com>